click on username to navigate to new component
